refactor array sequence window helpers
